Revert "Convert MenuModel to size_t."
This reverts commit c3d90e3544aac8ca2371d8ba423bdcdda3ef31af.

Reason for revert: This CL is possibly the cause of failure for 
quite a few of tests (app menu model unit tests on linux-chromeos-chrome builder, PWA browser tests, etc.), starting from https://ci.chromium.org/ui/p/chrome/builders/ci/linux-chromeos-chrome/23001/overview
also this is likely the cause of failed tast test: apps.LaunchHelpAppInBrowserMenu.stable in some builders, starting from: https://ci.chromium.org/ui/p/chrome/builders/ci/chromeos-betty-pi-arc-cfi-thin-lto-chrome/37592/overview
https://ci.chromium.org/p/chrome/builders/ci/chromeos-betty-pi-arc-chrome/23963

Original change's description:
> Convert MenuModel to size_t.
>
> Bug: 1292951
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3770735
> Commit-Queue: Peter Kasting <pkasting@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1026046}

Bug: 1292951
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3774644
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Owners-Override: Maggie Cai <mxcai@chromium.org>
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1026151}

+++ chrome/browser/ash/crosapi/test_controller_ash.h
@@ -120,7 +120,7 @@ class TestControllerAsh : public mojom::TestController,
   static void OnSelectContextMenuForShelfItem(
       SelectContextMenuForShelfItemCallback callback,
       const std::string& item_id,
-      size_t index,
+      uint32_t index,
       std::unique_ptr<ui::SimpleMenuModel> model);
 
   // Each call to EnterOverviewMode or ExitOverviewMode spawns a waiter for the
