+++ components/dbus/menu/menu_property_list.h
@@ -21,8 +21,7 @@ class MenuModel;
 
 // Computes properties for the menu item with index |i| in |menu|.
 COMPONENT_EXPORT(DBUS)
-MenuItemProperties ComputeMenuPropertiesForMenuItem(ui::MenuModel* menu,
-                                                    size_t i);
+MenuItemProperties ComputeMenuPropertiesForMenuItem(ui::MenuModel* menu, int i);
 
 // Given inputs |old_properties| and |new_properties|, computes outputs
 // |item_updated_props| and |item_removed_props| suitable for use in
