+++ ui/base/models/menu_model.h
@@ -55,52 +55,52 @@ class COMPONENT_EXPORT(UI_BASE) MenuModel
   virtual bool HasIcons() const = 0;
 
   // Returns the number of items in the menu.
-  virtual size_t GetItemCount() const = 0;
+  virtual int GetItemCount() const = 0;
 
   // Returns the type of item at the specified index.
-  virtual ItemType GetTypeAt(size_t index) const = 0;
+  virtual ItemType GetTypeAt(int index) const = 0;
 
   // Returns the separator type at the specified index.
-  virtual ui::MenuSeparatorType GetSeparatorTypeAt(size_t index) const = 0;
+  virtual ui::MenuSeparatorType GetSeparatorTypeAt(int index) const = 0;
 
   // Returns the command id of the item at the specified index.
-  virtual int GetCommandIdAt(size_t index) const = 0;
+  virtual int GetCommandIdAt(int index) const = 0;
 
   // Returns the label of the item at the specified index.
-  virtual std::u16string GetLabelAt(size_t index) const = 0;
+  virtual std::u16string GetLabelAt(int index) const = 0;
 
   // Returns the secondary label of the item at the specified index. Secondary
   // label is shown below the label.
-  virtual std::u16string GetSecondaryLabelAt(size_t index) const;
+  virtual std::u16string GetSecondaryLabelAt(int index) const;
 
   // Returns the minor text of the item at the specified index. The minor text
   // is rendered to the right of the label and using the font GetLabelFontAt().
-  virtual std::u16string GetMinorTextAt(size_t index) const;
+  virtual std::u16string GetMinorTextAt(int index) const;
 
   // Returns the minor icon of the item at the specified index. The minor icon
   // is rendered to the left of the minor text.
-  virtual ImageModel GetMinorIconAt(size_t index) const;
+  virtual ImageModel GetMinorIconAt(int index) const;
 
   // Returns true if the menu item (label/sublabel/icon) at the specified
   // index can change over the course of the menu's lifetime. If this function
   // returns true, the label, sublabel and icon of the menu item will be
   // updated each time the menu is shown.
-  virtual bool IsItemDynamicAt(size_t index) const = 0;
+  virtual bool IsItemDynamicAt(int index) const = 0;
 
   // Returns whether the menu item at the specified index has a user-set title,
   // in which case it should always be treated as plain text.
-  virtual bool MayHaveMnemonicsAt(size_t index) const;
+  virtual bool MayHaveMnemonicsAt(int index) const;
 
   // Returns the accessible name for the menu item at the specified index.
-  virtual std::u16string GetAccessibleNameAt(size_t index) const;
+  virtual std::u16string GetAccessibleNameAt(int index) const;
 
   // Returns the font list used for the label at the specified index.
   // If NULL, then the default font list should be used.
-  virtual const gfx::FontList* GetLabelFontListAt(size_t index) const;
+  virtual const gfx::FontList* GetLabelFontListAt(int index) const;
 
   // Gets the accelerator information for the specified index, returning true if
   // there is a shortcut accelerator for the item, false otherwise.
-  virtual bool GetAcceleratorAt(size_t index,
+  virtual bool GetAcceleratorAt(int index,
                                 ui::Accelerator* accelerator) const = 0;
 
   // Returns the checked state of the item at the specified index.
@@ -108,44 +108,44 @@ class COMPONENT_EXPORT(UI_BASE) MenuModel
 
   // Returns the id of the group of radio items that the item at the specified
   // index belongs to.
-  virtual int GetGroupIdAt(size_t index) const = 0;
+  virtual int GetGroupIdAt(int index) const = 0;
 
   // Gets the icon for the item at the specified index. ImageModel is empty if
   // there is no icon.
-  virtual ImageModel GetIconAt(size_t index) const = 0;
+  virtual ImageModel GetIconAt(int index) const = 0;
 
   // Returns the model for a menu item with a line of buttons at |index|.
-  virtual ButtonMenuItemModel* GetButtonMenuItemAt(size_t index) const = 0;
+  virtual ButtonMenuItemModel* GetButtonMenuItemAt(int index) const = 0;
 
   // Returns the enabled state of the item at the specified index.
-  virtual bool IsEnabledAt(size_t index) const = 0;
+  virtual bool IsEnabledAt(int index) const = 0;
 
   // Returns true if the menu item is visible.
-  virtual bool IsVisibleAt(size_t index) const;
+  virtual bool IsVisibleAt(int index) const;
 
   // Returns true if the item is rendered specially to draw attention
   // for in-product help.
-  virtual bool IsAlertedAt(size_t index) const;
+  virtual bool IsAlertedAt(int index) const;
 
   // Returns true if the menu item grants access to a new feature that we want
   // to show off to users (items marked as new will receive a "New" badge when
   // the appropriate flag is enabled).
-  virtual bool IsNewFeatureAt(size_t index) const;
+  virtual bool IsNewFeatureAt(int index) const;
 
   // Returns an application-window-unique identifier that can be used to track
   // the menu item irrespective of menu-specific command IDs.
-  virtual ElementIdentifier GetElementIdentifierAt(size_t index) const;
+  virtual ElementIdentifier GetElementIdentifierAt(int index) const;
 
   // Returns the model for the submenu at the specified index.
-  virtual MenuModel* GetSubmenuModelAt(size_t index) const = 0;
+  virtual MenuModel* GetSubmenuModelAt(int index) const = 0;
 
   // Called when the item at the specified index has been activated.
-  virtual void ActivatedAt(size_t index) = 0;
+  virtual void ActivatedAt(int index) = 0;
 
   // Called when the item has been activated with given event flags.
   // (for the case where the activation involves a navigation).
   // |event_flags| is a bit mask of ui::EventFlags.
-  virtual void ActivatedAt(size_t index, int event_flags);
+  virtual void ActivatedAt(int index, int event_flags);
 
   // Called when the menu is about to be shown.
   virtual void MenuWillShow() {}
@@ -169,7 +169,7 @@ class COMPONENT_EXPORT(UI_BASE) MenuModel
   // and specifies the model to start searching from.
   static bool GetModelAndIndexForCommandId(int command_id,
                                            MenuModel** model,
-                                           size_t* index);
+                                           int* index);
 
  private:
   // MenuModelDelegate. Weak. Could be null.
